[XML] Cleanup
